hippocampal/requirements.txt
# =============================================================================
# Ubik Hippocampal Node - Python Dependencies
# =============================================================================
# Install with: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Core MCP Framework
# -----------------------------------------------------------------------------
fastmcp==2.14.3
uvicorn==0.38.0

# -----------------------------------------------------------------------------
# Database Clients
# -----------------------------------------------------------------------------
# Neo4j graph database driver
neo4j==6.1.0

# ChromaDB vector database client
chromadb==1.4.1

# -----------------------------------------------------------------------------
# Data Validation
# -----------------------------------------------------------------------------
pydantic==2.12.5

# -----------------------------------------------------------------------------
# Embeddings
# -----------------------------------------------------------------------------
sentence-transformers==5.1.2

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dotenv==1.2.1
httpx==0.28.1

# -----------------------------------------------------------------------------
# Development Dependencies (optional)
# -----------------------------------------------------------------------------
# pytest>=8.0.0
# pytest-asyncio>=0.23.0
# pytest-cov>=4.1.0
